pyproject.toml
[project]
name = "01-fruit-ninja-refactoring"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bitsandbytes>=0.49.2",
    "diffusers>=0.37.1",
    "gsplat>=1.5.3",
    "hydra-core>=1.3.2",
    "mcubes>=0.1.6",
    "omegaconf>=2.3.0",
    "peft>=0.18.1",
    "plyfile>=1.1.3",
    "pydantic>=2.12.5",
    "pytorch-msssim>=1.0.0",
    "taichi>=1.7.4",
    "torch>=2.11.0",
    "torchvision>=0.26.0",
    "tqdm>=4.67.3",
    "transformers>=5.5.0",
    "wandb>=0.25.1",
]
